Change dollar attributes to decimals

+++ db/migrate/20180921182715_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
       t.text :comment
-      t.float :cost
+      t.decimal :cost
       t.references :shop, foreign_key: true
 
       t.timestamps
