+++ db/migrate/20180921190220_create_line_items.rb
@@ -2,7 +2,7 @@ class CreateLineItems < ActiveRecord::Migration[5.2]
   def change
     create_table :line_items do |t|
       t.integer :count
-      t.float :cost
+      t.decimal :cost
       t.belongs_to :order, foreign_key: true
       t.belongs_to :product, foreign_key: true
